Nouvels ajouts

+++ articles/supprimer.php
@@ -2,16 +2,18 @@
     session_start();
     include "../config/database.php";
 
-    $id = $row['id'];
+    if (isset($_GET['post_id'])) {
+        $id = $_GET['post_id'];
     
-    $sql = "DELETE FROM articles where id = ?";
+        $sql = "DELETE FROM articles where id = ?";
 
-    $result = $pdo->prepare($sql);
-    $result->execute(['$id']);
+        $result = $pdo->prepare($sql);
+        $result->execute(['$id']);
 
-    if(!$result){
-        echo"error! {$connexion->error}";
-    }else{
-        echo "Deleted successfully!";
+        if(!$result){
+            echo"error! {$pdo->error}";
+        }else{
+            echo "Deleted successfully!";
+        }
     }
 ?>
